remove an unused error, and move stuff around
